chore(build): 🧹 add clean-package to optimize package output and simplify exports
This fixes the need to build the package after installing it via a non
npm registry, but also requires anyone importing it via these methods to
be using some sort of typescript compiler.

What's done in the types export was required in order to have a coherent
way to import types both before and after the bundling is done. So right
now, all types are exported onto a single file, which is exported via
the /types export. This is also helpful in case some type was used in
both exports, since otherwise we were duplicating code.